JWT Authentication Feature
So yeah, while I was following some dud on youtube, i found this and i need this to proceed with my feature. This feature will not allow user to directly go to homepage without using credential.

Flaw: if a user can login as student or teacher, the user will be able to go the student or teacher homepage, even if the user is only a teacher or a student. In short, this feature was not yet fully distinguished.

Things to consider:
backend ->
npm install jsonwebtoken

package.json must be:
    "cookie-parser": "^1.4.5",